initial commit for flat maps

+++ nmcc/src/types.rs
@@ -11,16 +11,16 @@ impl<'a> std::error::Error for NmccError<'a> {}
 pub struct DecorReference {
     pub name: u32,
     pub texture: u32,
-    pub vertices: Vec<u32>,
-    pub uvs: Vec<u32>,
+    pub vertices: Vec<f32>,
+    pub uvs: Vec<f32>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct EntityReference {
     pub name: u32,
     pub texture: u32,
-    pub vertices: Vec<Vec<u32>>,
-    pub uvs: Vec<u32>,
+    pub vertices: Vec<Vec<f32>>,
+    pub uvs: Vec<f32>,
 }
 
 #[derive(Debug, PartialEq)]
@@ -29,15 +29,15 @@ pub struct EntityInstance {
     // names[index] == ogre, but also reference[index] == ${ogre_reference}
     pub index: Option<u32>,
     pub params: Vec<u32>, // indexes to [k,v,k,v,k,v] etc
-    pub location: [u32;3],
-    pub rotation: [u32;4],
-    pub scale: [u32;3],
+    pub location: [f32;3],
+    pub rotation: [f32;4],
+    pub scale: [f32;3],
 }
 
 #[derive(Debug, PartialEq)]
 pub struct DecorInstance {
     pub index: u32,
-    pub location: [u32;3],
-    pub rotation: [u32;4],
-    pub scale: [u32;3],
+    pub location: [f32;3],
+    pub rotation: [f32;4],
+    pub scale: [f32;3],
 }
